Include ingest and UI in publishing to traffic queue
